pyproject.toml
[project]
name = "flouri"
version = "0.1.0"
description = "AI-powered bash environment enhancement tool."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.22.0",
    "litellm>=1.81.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "click>=8.1.0",
    "rich>=13.7.0",
    "prompt-toolkit>=3.0.0",
    "pygments>=2.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
flouri = "flouri.ui.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["flouri"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
no_implicit_optional = false
ignore_missing_imports = true
follow_imports = "skip"
show_error_codes = true
allow_untyped_calls = true
allow_untyped_defs = true
exclude = [
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["flouri"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    # Hard-to-unit-test: interactive TUI, runner loop, prompt-toolkit–coupled completer, ROS2 completions
    "flouri/ui/tui.py",
    "flouri/runner/runner.py",
    "flouri/plugins/cd_completer.py",
    "flouri/completions/ros2.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
